+++ src/main/java/fontys/sem3/school/configuration/db/DatabaseDataInitializer.java
@@ -62,7 +62,7 @@ public class DatabaseDataInitializer {
                         .profilePictureUrl("https://firebasestorage.googleapis.com/v0/b/gourmet-eats.appspot.com/o/images%2Fuser%2Fjohndoe?alt=media&token=8615b91d-e8b4-49cf-92aa-a8b3a49cd798")
                         .role(Role.Seller)
                         .birthdate(LocalDate.of(1990, 1, 1))
-                        .balance(0L)
+                        .balance((double) 0)
                         .build(),
                 UserEntity.builder()
                         .name("johndoe1")
@@ -75,7 +75,7 @@ public class DatabaseDataInitializer {
                         .profilePictureUrl("https://firebasestorage.googleapis.com/v0/b/gourmet-eats.appspot.com/o/images%2Fuser%2Fjohndoe1.jpg?alt=media&token=75c89139-3982-4781-a557-0ff42452fa28")
                         .role(Role.Seller)
                         .birthdate(LocalDate.of(1995, 1, 1))
-                        .balance(0L)
+                        .balance((double) 0)
                         .build(),
                 UserEntity.builder()
                         .name("arielclarence")
@@ -88,7 +88,7 @@ public class DatabaseDataInitializer {
                         .profilePictureUrl("https://firebasestorage.googleapis.com/v0/b/gourmet-eats.appspot.com/o/images%2Fuser%2Fjohndoe?alt=media&token=8615b91d-e8b4-49cf-92aa-a8b3a49cd798")
                         .role(Role.Customer)
                         .birthdate(LocalDate.of(2002, 9, 19))
-                        .balance(0L)
+                        .balance((double) 0)
                         .build(),
                 UserEntity.builder()
                         .name("admin")
@@ -101,7 +101,7 @@ public class DatabaseDataInitializer {
                         .profilePictureUrl("https://firebasestorage.googleapis.com/v0/b/gourmet-eats.appspot.com/o/images%2Fuser%2Fadmin.png?alt=media&token=dba9b8e9-6ce0-4289-a922-f455583bfc90")
                         .role(Role.Admin)
                         .birthdate(LocalDate.of(2002, 9, 19))
-                        .balance(0L)
+                        .balance((double) 0)
                         .build()
         );
